Moved folders in to app folder - should fix some typical errors
