+++ Client/src/app/_models/joke.ts
