+++ Client/src/app/_models/user.ts
@@ -1,0 +1,7 @@
+export interface User {
+    id: string,
+    email: string,
+    userName: string,
+    role: string,
+    token: string
+}
