+++ Client/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { JokeDetailComponent } from 'src/components/jokes/joke-detail/joke-detail.component';
-import { JokeListComponent } from 'src/components/jokes/joke-list/joke-list.component';
+import { JokeDetailComponent } from './components/jokes/joke-detail/joke-detail.component';
+import { JokeListComponent } from './components/jokes/joke-list/joke-list.component';
 
 const routes: Routes = [
   { path: '', component: JokeListComponent },
