+++ Client/src/app/app.module.ts
@@ -4,11 +4,11 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 
 import { AppComponent } from '../app/app.component';
-import { JokeDetailComponent } from '../components/jokes/joke-detail/joke-detail.component';
-import { NavbarComponent } from '../components/navbar/navbar.component';
 import { AppRoutingModule } from './app-routing.module';
-import { JokeCardComponent } from 'src/components/jokes/joke-card/joke-card.component';
-import { JokeListComponent } from 'src/components/jokes/joke-list/joke-list.component';
+import { NavbarComponent } from './components/navbar/navbar.component';
+import { JokeCardComponent } from './components/jokes/joke-card/joke-card.component';
+import { JokeDetailComponent } from './components/jokes/joke-detail/joke-detail.component';
+import { JokeListComponent } from './components/jokes/joke-list/joke-list.component';
 
 @NgModule({
   imports: [
