+++ Client/src/app/components/jokes/joke-card/joke-card.component.html
