+++ Client/src/app/components/jokes/joke-detail/joke-detail.component.html
