+++ Client/src/app/components/jokes/joke-list/joke-list.component.html
